preformance fix

+++ Business.Shared/DTOs/ProductIngredientAddUpdateDTO.cs
@@ -18,8 +18,8 @@ namespace Business.Shared.DTOs
         public int ProductId { get; set; }
 
         [Required]
-        [Range(0, double.MaxValue)]
-        public double Quantity { get; set; }
+        [Range(0, float.MaxValue)]
+        public float Quantity { get; set; }
 
         [Required]
         [StringLength(50)]
