+++ Business.Shared/DTOs/PurchasesDTO.cs
@@ -31,7 +31,7 @@ namespace Business.Shared.DTOs
  
 
         [Required]
-        public int Quantity { get; set; } // الكمية
+        public float Quantity { get; set; } // الكمية
 
         [Required]
         [StringLength(50)] // تحديد الحد الأقصى لعدد الحروف
