+++ Business.Shared/DTOs/StockDTO.cs
@@ -15,7 +15,7 @@ namespace Business.Shared.DTOs
         public string ItemName { get; set; }
 
         [Required]
-        public int AvailableQuantity { get; set; }
+        public float AvailableQuantity { get; set; }
 
         [StringLength(50, ErrorMessage = "lenght max 50")]
         public string UnitOfMeasure { get; set; }
